requirements.txt
apturl==0.5.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
boto3==1.26.98
botocore==1.29.98
branca==0.6.0
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
command-not-found==0.3
contourpy==1.0.7
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
descartes==1.1.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
executing==1.2.0
fasteners==0.14.1
Fiona==1.9.2
folium==0.14.0
fonttools==4.39.2
future==0.18.2
geopandas==0.12.2
httplib2==0.14.0
idna==2.8
imageio==2.26.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipython==8.11.0
ipywidgets==8.0.5
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyterlab-widgets==3.0.6
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
monotonic==1.5
munch==2.5.0
netifaces==0.10.4
networkx==3.0
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
osmnx==1.2.2
OWSLib==0.28.1
packaging==23.0
pandas==1.5.3
paramiko==2.6.0
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
protobuf==3.6.1
pure-eval==0.2.2
pycairo==1.16.2
pycups==1.9.73
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyproj==3.4.1
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.7.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.2
requests-unixsocket==0.2.0
Rtree==1.0.1
s3transfer==0.6.0
SecretStorage==2.3.1
Shapely==1.8.5.post1
simplejson==3.16.0
six==1.14.0
stack-data==0.6.2
systemd-python==234
traitlets==5.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.6
widgetsnbextension==4.0.6
xkit==0.0.0
zipp==3.15.0
